extra_scripts/requirements.txt
gradio==4.19.2
numpy==1.26.4
scikit-learn==1.4.0
google-generativeai==0.3.2
pandas==2.2.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
torch==2.2.1 